# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RunSafe API — SMB Operations & Compliance SaaS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /compliance/items:
    get:
      operationId: getComplianceItems
      tags: [compliance]
      summary: List compliance items for the workspace
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of compliance items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComplianceItem"
    post:
      operationId: createComplianceItem
      tags: [compliance]
      summary: Create a compliance item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComplianceItemBody"
      responses:
        "201":
          description: Created compliance item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceItem"
  /compliance/items/{itemId}:
    patch:
      operationId: updateComplianceItem
      tags: [compliance]
      summary: Update a compliance item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComplianceItemBody"
      responses:
        "200":
          description: Updated compliance item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceItem"
    delete:
      operationId: deleteComplianceItem
      tags: [compliance]
      summary: Delete a compliance item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /compliance/items/{itemId}/complete:
    post:
      operationId: completeComplianceItem
      tags: [compliance]
      summary: Mark a compliance item as completed (audit log entry)
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteComplianceItemBody"
      responses:
        "201":
          description: Completion recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceCompletion"
...
components:
  schemas:
...
    ComplianceItem:
      type: object
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [employment, health_safety, data_privacy, licensing, other]
        recurrence:
          type: string
          enum: [one_time, monthly, quarterly, annually]
        dueDate:
          type: ["string", "null"]
          format: date
        lastCompletedAt:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [pending, upcoming, overdue, completed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - workspaceId
        - title
        - category
        - recurrence
        - status
        - createdAt
    CreateComplianceItemBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [employment, health_safety, data_privacy, licensing, other]
        recurrence:
          type: string
          enum: [one_time, monthly, quarterly, annually]
        dueDate:
          type: string
          format: date
      required:
        - title
        - category
        - recurrence
    UpdateComplianceItemBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [employment, health_safety, data_privacy, licensing, other]
        recurrence:
          type: string
          enum: [one_time, monthly, quarterly, annually]
        dueDate:
          type: string
          format: date
    CompleteComplianceItemBody:
      type: object
      properties:
        notes:
          type: string
        completedAt:
          type: string
          format: date-time
    ComplianceCompletion:
      type: object
      properties:
        id:
          type: integer
        complianceItemId:
          type: integer
        complianceItemTitle:
          type: string
        completedByClerkId:
          type: string
        completedByName:
          type: string
        notes:
          type: ["string", "null"]
        completedAt:
          type: string
          format: date-time
      required:
        - id
        - complianceItemId
        - complianceItemTitle
        - completedByClerkId
        - completedByName
        - completedAt

    # ─── DASHBOARD ───────────────────────────────────────────────
...